+++ src/Monsters/Tables/Level1.js
@@ -10,7 +10,7 @@ import {CaveSquig} from "../Races/Monsters/CaveSquigs";
 import {Centaur} from "../Races/Monsters/Centaurs";
 import {GiantBat} from "../Races/Monsters/GiantBat";
 import {GiantRat} from "../Races/Monsters/GiantRats";
-import {GiantSpider} from "../Races/Monsters/GiantSpider";
+import {GiantSpider} from "../Races/Monsters/GiantSpiders";
 import {Minotaur} from "../Races/Monsters/Minotaurs";
 import {Ogre} from "../Races/Monsters/Ogre";
